+++ {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/static_source/css/base.scss
@@ -27,7 +27,6 @@
  *
  */
 
-@import "@/../templates/components/links/links.scss";
 @import "@/../templates/components/messages.scss";
