+++ {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/templates/base.jinja
@@ -17,6 +17,7 @@
         {{ vite_hmr_client() }}
         {{ vite_asset('css/styles.js') }}
         {{ vite_asset('js/components.ts') }}
+        {{ vite_asset('js/links.ts') }}
         {{ vite_asset('js/main.ts') }}
         {{ django_htmx_script() }}
         <title>{% block title required %}{% endblock title %} - Lightmatter</title>
