+++ {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/templates/components/flyout/flyout.jinja
@@ -1,6 +1,7 @@
 {% from 'components/button/button.jinja' import button %}
 {% from 'components/util.jinja' import attributes %}
 {% set default_button %}
+{# see button.jinja for more details about include_x_data #}
 {{ button(include_x_data=False, attrs={"x-menu:button": True})}}
 {% endset %}
 {% macro flyout(button=default_button, menu_style="absolute right-0 w-48 mt-2 z-10 origin-top-right bg-white border border-gray-200 divide-y divide-gray-100 rounded-md shadow-md py-1 outline-none", menu_attrs={}) %}
