+++ {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/templates/header/end.jinja
@@ -16,6 +16,7 @@
 {% endset %}
 {% set menu_style = "origin-top-right absolute right-0 mt-2 w-48 py-1 bg-white focus:outline-none" %}
 <div class="hidden md:ml-4 md:flex-shrink-0 md:flex md:items-center">
+    {# see button.jinja for more details about include_x_data #}
     {% call flyout(button(button_for_menu, include_x_data=False, attrs={"x-menu:button": True}), menu_style) %}
         <div role="menu"
             aria-orientation="vertical"
